Rename DolbyIOModule to DolbyIO
