+++ DolbyIO/DolbyIO.uplugin
@@ -15,7 +15,7 @@
 	"Installed": false,
 	"Modules": [
 		{
-			"Name": "DolbyIOModule",
+			"Name": "DolbyIO",
 			"Type": "Runtime",
 			"LoadingPhase": "PreDefault",
 			"WhitelistPlatforms": [ "Win64", "Mac" ]
