+++ DolbyIO/Source/DolbyIO/DolbyIO.Build.cs
@@ -3,9 +3,9 @@
 using System.IO;
 using UnrealBuildTool;
 
-public class DolbyIOModule : ModuleRules
+public class DolbyIO : ModuleRules
 {
-	public DolbyIOModule(ReadOnlyTargetRules Target) : base(Target)
+	public DolbyIO(ReadOnlyTargetRules Target) : base(Target)
 	{
 		PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
