+++ DolbyIO/Source/DolbyIO/Public/DolbyIOParticipantInfo.h
@@ -7,7 +7,7 @@
 /* Contains the current status of a conference participant and information whether the participant's audio is enabled.
  */
 USTRUCT(BlueprintType, Category = "Dolby.io Comms", DisplayName = "Dolby.io Participant Info")
-struct DOLBYIOMODULE_API FDolbyIOParticipantInfo
+struct DOLBYIO_API FDolbyIOParticipantInfo
 {
 	GENERATED_BODY()
